# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Curriculum Builder API",
    "description": "AI-Assisted Instructional Design API. Generate structured training curricula with learning objectives, module breakdowns, activities, and assessments. Targets Docebo, Workday Learning, SuccessFactors, SCORM 2004 / xAPI.",
    "version": "1.0.0",
    "contact": {
      "name": "Curriculum Builder"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "paths": {
    "/api/generate": {
      "post": {
        "summary": "Generate curriculum",
        "description": "Generates a complete curriculum structure using Claude AI based on topic, audience, format, and duration parameters.",
        "operationId": "generateCurriculum",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt"],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "The full prompt sent to Claude (includes topic, audience, format, duration, and JSON schema instructions)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful generation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "type": { "type": "string" },
                    "role": { "type": "string" },
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": { "type": "string", "enum": ["text"] },
                          "text": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - no prompt provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "No prompt provided" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error - API key not configured or Claude API error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}